backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
Pillow